requirements.txt
Django==5.0.2
django-admin==2.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
requests==2.25.1
python-dotenv
